Move `Theme` type to `iced_core`
