+++ core/Cargo.toml
@@ -19,9 +19,8 @@ smol_str.workspace = true
 thiserror.workspace = true
 web-time.workspace = true
 xxhash-rust.workspace = true
-
 palette.workspace = true
-palette.optional = true
+once_cell.workspace = true
 
 [target.'cfg(windows)'.dependencies]
 raw-window-handle.workspace = true
