+++ core/src/lib.rs
@@ -30,6 +30,7 @@ pub mod overlay;
 pub mod renderer;
 pub mod svg;
 pub mod text;
+pub mod theme;
 pub mod time;
 pub mod touch;
 pub mod widget;
@@ -76,6 +77,7 @@ pub use shadow::Shadow;
 pub use shell::Shell;
 pub use size::Size;
 pub use text::Text;
+pub use theme::Theme;
 pub use transformation::Transformation;
 pub use vector::Vector;
 pub use widget::Widget;
