+++ examples/color_palette/Cargo.toml
@@ -7,6 +7,6 @@ publish = false
 
 [dependencies]
 iced.workspace = true
-iced.features = ["canvas", "palette"]
+iced.features = ["canvas"]
 
 palette.workspace = true
