+++ widget/Cargo.toml
@@ -25,7 +25,6 @@ wgpu = ["iced_renderer/wgpu"]
 [dependencies]
 iced_renderer.workspace = true
 iced_runtime.workspace = true
-iced_style.workspace = true
 
 num-traits.workspace = true
 thiserror.workspace = true
